Encode BIGINT as string in JSON and fix bug in aggregation forward key

+++ push/exec/exec.go
@@ -31,7 +31,7 @@ func NewExecutionContext(writeBatch *cluster.WriteBatch, enableDuplicateDetectio
 type ExecutionContext struct {
 	WriteBatch               *cluster.WriteBatch
 	RemoteBatches            map[uint64]*cluster.WriteBatch
-	BatchSequence            uint64
+	BatchSequence            uint32
 	EnableDuplicateDetection bool
 }
